+++ sopa/_constants.py
@@ -16,6 +16,7 @@ class SopaKeys:
     INSTANCE_KEY = "cell_id"
     BAYSOR_DEFAULT_CELL_KEY = "cell"
     BAYSOR_AREA_OBS = "baysor_area"
+    AREA_OBS = "area"
 
     Z_SCORES = "z_scores"