+++ sopa/segmentation/stainings.py
@@ -127,7 +127,7 @@ class StainingSegmentation:
         gdf = gpd.GeoDataFrame(geometry=cells)
 
         patch_dir: Path = Path(patch_dir)
-        patch_dir.mkdir(exist_ok=True)
+        patch_dir.mkdir(parents=True, exist_ok=True)
         patch_file = patch_dir / f"{patch_index}.parquet"
 
         gdf.to_parquet(patch_file)