+++ sopa/spatial/distance.py
@@ -99,4 +99,62 @@ def mean_distance(
         df_distances.replace(0, np.nan, inplace=True)
 
     df_distances[group_key] = adata.obs[group_key]
-    return df_distances.groupby(group_key, observed=False).mean()
+    df_distances = df_distances.groupby(group_key, observed=False).mean()
+    df_distances.columns.name = target_group_key
+
+    return df_distances
+
+
+def prepare_network(
+    adata: AnnData,
+    cell_type_key: str,
+    niche_key: str,
+    clip_weight: float = 3,
+    node_colors: tuple[str] = ("#5c7dc4", "#f05541"),
+    node_sizes: float = (1.3, 5),
+) -> tuple[pd.DataFrame, dict, dict, dict]:
+    """Create a dataframe representing weights between cell-types and/or niches.
+    This can be later use to plot a cell-type/niche represention of a whole slide
+    using the netgraph library.
+
+    Args:
+        adata: An `AnnData` object
+        cell_type_key: Key of `adata.obs` containing the cell types
+        niche_key: Key of `adata.obs` containing the niches
+        clip_weight: Maximum weight
+        node_colors: Tuple of (cell-type color, niche color)
+        node_sizes: Tuple of (cell-type size, niche size)
+
+    Returns:
+        A DataFrame of weights between cell-types and/or niches, and three dict for netgraph display
+    """
+    node_color, node_size, node_shape = {}, {}, {}
+
+    log.info("Computing all distances for the 4 pairs of categories")
+    weights = mean_distance(adata, cell_type_key)
+    top_right = mean_distance(adata, cell_type_key, niche_key)
+    bottom_left = mean_distance(adata, niche_key, cell_type_key)
+    bottom_right = mean_distance(adata, niche_key, niche_key)
+
+    for pop in weights.index:
+        node_color[pop] = node_colors[0]
+        node_size[pop] = node_sizes[0]
+        node_shape[pop] = "o"
+
+    for niche in bottom_right.index:
+        node_color[niche] = node_colors[1]
+        node_size[niche] = node_sizes[1]
+        node_shape[niche] = "h"
+
+    # assemble dataframe per-block
+    bottom_left[bottom_right.columns] = bottom_right
+    weights[top_right.columns] = top_right
+    weights = pd.concat([weights, bottom_left], axis=0).copy()
+
+    # convert distances to symmetric weights
+    weights = 1 / weights
+    np.fill_diagonal(weights.values, 0)
+    weights = weights.clip(0, clip_weight)
+    weights = (weights.T + weights) / 2
+
+    return weights, node_color, node_size, node_shape