+++ CHANGELOG.md
@@ -1,11 +1,17 @@
-## [1.0.x] - tbd
+## [1.0.2] - 2024-01-15
 
 ### Fix
 - When geometries are `GeometryCollection`, convert them back to Polygons (#11)
-- Snakemake pipeline fixed when providing `min_area` parameter to Baysor
+- Give `min_area` parameter to the right Baysor function in snakemake
 
 ### Added
+- API tutorial
+- `sopa.spatial` tutorial
 - Docstrings for the snakemake pipeline utils
+- Show right micron scale in the Xenium Explorer
+
+### Changed
+- `sopa.stats` is now called `sopa.spatial`
 
 ## [1.0.1] - 2024-01-10